+++ src/super/i-static-page/modules/interface.js
@@ -23,24 +23,42 @@ exports.LibSource = LibSource;
  *
  * * src - relative path to a file to load, i.e. without referencing to /node_modules, etc.
  * * [source='lib'] - source type of the library, i.e. where the library is stored
- * * [inline=false] - if true, the library will be placed as text into a script tag
- * * [defer=true] - if true, the script is declared with the "defer" attribute
- * * [async=false] - if true, the script is declared with the "async" attribute
- * * [module=false] - if true, the script is declared with the "module" attribute
- * * [load=true] - if false, the script won't be automatically loaded with a page
+ * * [inline=false] - if true, the library is placed as a text
+ * * [defer=true] - if true, the library is declared with the "defer" attribute
+ * * [load=true] - if false, the library won't be automatically loaded with a page
+ * * [attrs] - dictionary with additional attributes
  *
  * @typedef {{
  *   src: string,
  *   source?: LibSource,
  *   inline?: boolean,
- *   defer?: boolean,
- *   async?: boolean,
- *   module?: boolean,
- *   load?: boolean
+ *   load?: boolean,
+ *   attrs?: Object
  * }}
  */
 const Lib = {};
 exports.Lib = Lib;
+
+/**
+ * Parameters of an initialized script library:
+ *
+ * * src - path to a file to load
+ * * [documentWrite] - if true, the function returns JS code to load the library by using document.write
+ * * [staticAttrs] - string with additional attributes
+ *
+ * @see Lib
+ * @typedef {{
+ *   src: string,
+ *   inline?: boolean,
+ *   defer?: boolean,
+ *   load?: boolean,
+ *   documentWrite?: boolean,
+ *   attrs?: Object,
+ *   staticAttrs?: string
+ * }}
+ */
+const InitializedLib = {};
+exports.InitializedLib = InitializedLib;
 
 /**
  * Map of script libraries to require:
@@ -56,18 +74,40 @@ exports.Libs = Libs;
  *
  * * src - relative path to a file to load, i.e. without referencing to /node_modules, etc.
  * * [source='lib'] - source type of the library, i.e. where the library is stored
- * * [inline=false] - if true, the library will be placed as text into a style tag
- * * [defer=true] - if true, the style will be loaded only after loading of the whole page
+ * * [inline=false] - if true, the library is placed as text into a style tag
+ * * [defer=true] - if true, the library is loaded only after loading of the whole page
+ * * [attrs] - dictionary with additional attributes
  *
  * @typedef {{
  *   src: string,
  *   source?: LibSource,
  *   inline?: boolean,
  *   defer?: boolean
+ *   attrs?: Object
  * }}
  */
 const StyleLib = {};
 exports.StyleLib = StyleLib;
+
+/**
+ * Parameters of an initialized style library:
+ *
+ * * src - path to a file to load
+ * * [documentWrite] - if true, the function returns JS code to load the library by using document.write
+ * * [staticAttrs] - string with additional attributes
+ *
+ * @see StyleLib
+ * @typedef {{
+ *   src: string,
+ *   inline?: boolean,
+ *   defer?: boolean,,
+ *   documentWrite?: boolean,
+ *   attrs?: Object,
+ *   staticAttrs?: string
+ * }}
+ */
+const InitializedStyleLib = {};
+exports.InitializedStyleLib = InitializedStyleLib;
 
 /**
  * Map of style libraries to require:
@@ -77,3 +117,46 @@ exports.StyleLib = StyleLib;
  */
 const StyleLibs = new Map();
 exports.StyleLibs = StyleLibs;
+
+/**
+ * Parameters of a link:
+ *
+ * * src - relative path to a file to load, i.e. without referencing to /node_modules, etc.
+ * * [source='lib'] - source type of the library, i.e. where the library is stored
+ * * [attrs] - dictionary with additional attributes
+ *
+ * @typedef {{
+ *   src: string,
+ *   source?: LibSource,
+ *   attrs?: Object
+ * }}
+ */
+const Link = {};
+exports.Link = Link;
+
+/**
+ * Parameters of an initialized link:
+ *
+ * * src - path to a file to load
+ * * [documentWrite] - if true, the function returns JS code to load the library by using document.write
+ * * [staticAttrs] - string with additional attributes
+ *
+ * @see Link
+ * @typedef {{
+ *   src: string,
+ *   documentWrite?: boolean,
+ *   attrs?: Object,
+ *   staticAttrs?: string
+ * }}
+ */
+const InitializedLink = {};
+exports.InitializedLink = InitializedLink;
+
+/**
+ * Map of links to require:
+ * the value can be declared as a string (relative path to a file to load) or object with parameters
+ *
+ * @typedef {Map<string, (string|StyleLib)>}
+ */
+const Links = new Map();
+exports.Links = Links;